methodologies/ensemble/requirements.txt
anthropic>=0.40
openai>=1.50
python-dotenv>=1.0
pytest>=8.0
